# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Overthrone game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teams/{id}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Get a specific team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          description: Team not found
...
components:
  schemas:
...
    Team:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        hp:
          type: integer
        ap:
          type: integer
        members:
          type: array
          items:
            type: string
        isEliminated:
          type: boolean
        allianceId:
          type: ["integer", "null"]
        activeTaskId:
          type: ["integer", "null"]
        isAdmin:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - hp
        - ap
        - members
        - isEliminated
        - allianceId
        - activeTaskId
        - isAdmin
        - createdAt
...